requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
Flask==3.1.2
frozenlist==1.8.0
google-auth==2.41.1
google-auth-oauthlib==1.2.3
gspread==6.2.1
httplib2==0.31.0
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
multidict==6.7.0
oauth2client==4.1.3
oauthlib==3.3.1
propcache==0.4.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
pyparsing==3.2.5
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
twilio==9.8.6
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.4
yarl==1.22.0
zipp==3.23.0
